♻️ B2XY00 code: unify manifest scripts
