+++ packages/recipes/package.json
@@ -38,7 +38,7 @@
   },
   "scripts": {
     "clean": "node -e \"require('node:fs').rmSync('dist',{recursive:true,force:true}); require('node:fs').rmSync('tsconfig.tsbuildinfo',{force:true});\"",
-    "build": "tsc -b && node ../../scripts/write-build-manifest.mjs .",
+    "build": "tsc -b && node ../../scripts/manifest.mjs build .",
     "typecheck": "tsc -b",
     "prepack": "bun run clean && bun run build",
     "prepublishOnly": "node ../../scripts/enforce-github-publish.mjs && npm run prepack",
